services/planning/pyproject.toml
[project]
name = "productivity-agent-planning"
version = "0.1.0"
description = "Planning backend for the personal productivity agent"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.0.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "pydantic>=2.8.0",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
planning-init-db = "productivity_agent.init_db:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
